+++ fabfed/controller/chi/chi_slice.py
@@ -26,10 +26,10 @@ import logging
 
 from tabulate import tabulate
 
-from mobius.controller.chi.chi_network import Network
-from mobius.controller.chi.chi_node import Node
-from mobius.controller.util.config import Config
-from mobius.models import AbstractSlice
+from fabfed.controller.chi.chi_network import Network
+from fabfed.controller.chi.chi_node import Node
+from fabfed.controller.util.config import Config
+from fabfed.models import AbstractSlice
 
 
 class Slice(AbstractSlice):
