+++ fabfed/controller/fabric/fabric_client.py
@@ -27,10 +27,10 @@ from typing import List, Dict
 
 from fabrictestbed_extensions.fablib.fablib import fablib
 from fabrictestbed_extensions.fablib.resources import Resources
-from mobius.controller.api.api_client import ApiClient
-from mobius.controller.fabric.fabric_slice import FabricSlice
-from mobius.controller.util.config import Config
-from mobius.models import AbstractResourceListener
+from fabfed.controller.api.api_client import ApiClient
+from fabfed.controller.fabric.fabric_slice import FabricSlice
+from fabfed.controller.util.config import Config
+from fabfed.models import AbstractResourceListener
 
 
 class FabricClient(ApiClient, AbstractResourceListener):
