+++ fabfed/test/new_controller_test.py
@@ -1,4 +1,4 @@
-from mobius.controller.controller import Controller
+from fabfed.controller.controller import Controller
 
 if __name__ == "__main__":
     controller = Controller(config_file_location="new_config.yml")
